build: Add CMakeLists.txt and ignore build directory
This commit adds a CMakeLists.txt file to the project root directory. The
file sets up the project, enables language support for C++, and sets the
C++ standard to 17. It also includes compiler flags for Clang, GNU, and MSVC.

The commit also adds a .gitignore rule that ignores the build/ directory.
This is necessary because it will contain all of our generated files from
building with cmake.

Additionally, this commit introduces `CPM.cmake` as an external dependency.
It downloads version 0.38.1 of `CPM.cmake` if it doesn't already exist in
the cache or at `${CMAKE_BINARY_DIR}/cmake`.