feat: add helper function to generate RealOp_to_OID table
This commit adds a new helper function called `Generate_RealOp_to_OID()` to the `Opcode_table_generator` class. This function generates a lookup table that maps real opcodes to their corresponding OIDs. The implementation of this function involves finding the OID from OID_to_OI and enumerating it.

The purpose of this change is to prepare for future SDK development where we will need this lookup table.