feat: add debug function to get OID_to_OI map
This commit adds a new debug function `GetOID_to_OI()` to the `Opcode_table_generator` class. This function returns the map of Object ID (OID) to Opcode Index (OI). The implementation is added in both header and source files. 

The purpose of this change is to make it easier for developers to debug issues related to OIDs and OIs during runtime by providing access to the mapping between them.