build: update retdec installation process
This commit updates the installation process of retdec in CMakeLists.txt. The "DEBUG" macro was causing a conflict with some `enum` value names in retdec, so it was undefined before including the headers and then recovered after that. Also, we changed DCMAKE_INSTALL_PREFIX to our local directory instead of /usr/local/include/capstone.

CPMAddPackage is used to add gh:25077667/retdec@5.1 as a dependency package for this project.

The changes will help us avoid conflicts and improve the build process.